requirements.txt
jsonschema
oslo.config>=2.2.0
pbr
PyYAML
requests
